Watch mount for FAN_CREATE event
When using recursive inode watches, try to setup a mount mark for
FAN_CREATE events in case we have permissions on the mount in addition
to the recursive fanotify inode watches.

This will address the lost 'create' events when creating a directory tree
with only recursive inode watches.  For example, when executing command
'mkdir -p A/B/C', the create events for B and C will be reported.
If watcher program has capabilities on the mount userns it will also be
able to resolve those events to paths and add inode watches on the
created directories.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>