Support watching a mount with fanotify
With fsnotify{wait,watch} --mount, try to setup a mount mark.
User may be able to do that with an idmapped mount inside userns,
where watching the whole filesystem is not allowed.

Note that some events (e.g. create,delete) are not supported with
a mount mark, so this mode is limited to users that want the supported
events (e.g. open,close).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>