Support watching a filesystem with fanotify inside user namespace
Using --filesystem to watch a filesystem inside non-root userns is
allowed for user 0 of the user namespace, but only for filesytsems
that were mounted by this user (e.g. tmpfs) in its mount namespace.

In this case, it is required to use the O_DIRECTORY flag with
open_by_handle_at(2) to allow opening a directory file by user 0
of a non-root user namespace.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>