Index mount fd for client so it can resolve fid to path
When running fsnotifywait --server socket -S /path/a/ /path/b/
and fsnotifywait --client socket -m, the client received events with
fid of parent but cannot resolve those fid to path.

User needs to run fsnotifywait --client socket -m /path/a/ /path/b/
to prove that it has access to those mounts and to index the mount fds
for those mounts so that it could resolve fid to path.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>